Move auth files to /auth
