+++ lib/url-helper.js
@@ -1,6 +1,6 @@
 export default function createLoginUrl(redirectTo) {
 	if (redirectTo) {
-		return `/api/login?redirectTo=${encodeURIComponent(redirectTo)}`;
+		return `/api/auth/login?redirectTo=${encodeURIComponent(redirectTo)}`;
 	}
-	return `/api/login`;
+	return `/api/auth/login`;
 }
