+++ next.config.js
@@ -21,7 +21,7 @@ module.exports = {
 		AUTH0_CLIENT_SECRET: process.env.AUTH0_CLIENT_SECRET,
 		AUTH0_SCOPE: process.env.AUTH0_SCOPE,
 		REDIRECT_URI:
-			process.env.REDIRECT_URI || "http://localhost:3000/api/callback",
+			process.env.REDIRECT_URI || "http://localhost:3000/api/auth/callback",
 		POST_LOGOUT_REDIRECT_URI:
 			process.env.POST_LOGOUT_REDIRECT_URI || "http://localhost:3000/",
 		SESSION_COOKIE_SECRET: process.env.SESSION_COOKIE_SECRET,
