+++ pages/api/auth/callback.js
@@ -1,4 +1,4 @@
-import auth0 from "../../lib/auth0";
+import auth0 from "../../../lib/auth0";
 
 export default async function callback(req, res) {
 	try {
